Permitindo parametros float nas requisicoes

+++ src/controllers/validator/requestvalidator.py
@@ -6,7 +6,7 @@ class RequestValidator():
     def validateConvertRequestArgs(args):
         from_currency = args.get('from',type=str)
         to_currency = args.get('to',type=str)
-        amount = args.get('amount',type=int)
+        amount = args.get('amount',type=float)
         model = ConvertRequestModel(to_currency,from_currency,amount)
         if(model.isDataNull()):
             raise InvalidParametersException()
@@ -21,7 +21,7 @@ class RequestValidator():
 
     def validateCurrencyPutRequestArgs(args):
         name = args.get('name',type=str)
-        value = args.get('value',type=int)
+        value = args.get('value',type=float)
         model = CurrencyPutRequestModel(name,value)
         if(model.isDataNull()):
             raise InvalidParametersException()
